ExtremeWorldBackend/src/test/java/extremeworld/controller/ReservationControllerTest.java
package extremeworld.controller;

import Util.ReservationDtoUtil;
import Util.ResortDtoUtil;
import com.fasterxml.jackson.databind.ObjectMapper;
import extremeworld.dto.ReservationDTO;
import extremeworld.dto.ResortDTO;
import extremeworld.service.ReservationService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(controllers = ReservationController.class)
class ReservationControllerTest {

    @MockBean
    ReservationService reservationService;

    @Autowired
    MockMvc mockMvc;

    @Autowired
    ObjectMapper objectMapper;

    @Test
    void create() throws Exception{
        ReservationDTO dto = ReservationDtoUtil.generateReservationDTO("2022-01-01");
        when(reservationService.create(any())).thenReturn(dto);

        MvcResult result = mockMvc.perform(post("/reservations/create")
                .content(objectMapper.writeValueAsString(dto))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andReturn();

        assertThat(result.getResponse().getContentAsString()).isEqualTo(objectMapper.writeValueAsString(dto));
    }
}